requirements.txt
cython==0.29.20
pandas==1.0.5
requests==2.24.0
certifi==2020.6.20
cryptography==2.9.2
chardet==3.0.4
idna==2.9
azure-core==1.6.0
azure-common==1.1.25
azure-nspkg==3.0.2
azure-functions==1.2.1
azure-storage-common==2.1.0
azure-storage-nspkg==3.1.0
azure-storage-blob==12.3.2
azure-storage-file-share==12.1.2
redis==3.5.3
azure-cosmos==4.1.0
PyYAML==5.3.1
jinja2==2.11.2
pytz==2020.01
